# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NeverGuess API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}/share:
    post:
      tags: [Audits]
      operationId: createReportShareLink
      summary: Generate or return a public share slug for a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share link generated or already existed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareLinkResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Audits]
      operationId: revokeReportShareLink
      summary: Revoke an existing share link for a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share link revoked (or already absent).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeShareLinkResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /r/{slug}:
    get:
      tags: [Audits]
      operationId: getPublicReport
      summary: Get a publicly shared report by slug (no auth)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public report returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicReport'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    RiskSeverity:
      type: string
      enum: [critical, high, medium, low]
    RiskEvidence:
      type: object
      description: |
        A concrete pointer into the ingested repo backing a risky assumption.
        `verified` is stamped server-side after cross-checking the cited path
        against the ingested file tree.
      required: [path]
      properties:
        path:
          type: string
        quote:
          type: [string, 'null']
        verified:
          type: boolean
    RiskyAssumption:
      type: object
      required: [title, severity, detail]
      properties:
        title:
          type: string
        severity:
          $ref: '#/components/schemas/RiskSeverity'
        detail:
          type: string
        evidence:
          type: array
          description: Absent on reports generated before citations shipped.
          items:
            $ref: '#/components/schemas/RiskEvidence'
    AffectedArea:
      type: object
      description: |
        Blast radius entry — a file or directory the requested change is
        expected to touch. Paths are model-proposed; `create` entries may
        not exist in the tree yet.
      required: [path, reason, action]
      properties:
        path:
          type: string
        reason:
          type: string
        action:
          type: string
          enum: [modify, create, read]
    AcceptanceCriterion:
      type: object
      required: [title, detail]
      properties:
        title:
          type: string
        detail:
          type: string
    PromptPack:
      type: object
      required: [replit, cursor, copilot, claudeCode, codex]
      properties:
        replit:
          type: string
        cursor:
          type: string
        copilot:
          type: string
        claudeCode:
          type: string
        codex:
          type: string
    Verdict:
      type: string
      enum: [safe, caution, block]
    TestPlan:
      type: object
      required: [vitest, playwright]
      properties:
        vitest:
          type: string
        playwright:
          type: string
    SmokeCheckStatus:
      type: string
      enum: [pass, fail, warn, skip]
    SmokeCheck:
      type: object
      required: [id, label, status, detail]
      properties:
        id:
          type: string
        label:
          type: string
        status:
          $ref: '#/components/schemas/SmokeCheckStatus'
        metric:
          type: string
        detail:
          type: string
    SmokeTestResults:
      type: object
      required: [ranAt, url, skipped, checks]
      properties:
        ranAt:
          type: string
          format: date-time
        url:
          type: string
        skipped:
          type: boolean
        skipReason:
          type: string
        checks:
          type: array
          items:
            $ref: '#/components/schemas/SmokeCheck'
...
    AnalysisUsage:
      type: object
      description: Model response receipt captured from the analysis provider.
      required: [model, requestedModel, reasoningEffort, gateway, generatedAt]
      properties:
        model:
          type: string
        requestedModel:
          type: string
        reasoningEffort:
          type: string
        gateway:
          type: string
          enum: [replit-ai-integrations, openrouter-direct]
        requestId:
          type: [string, 'null']
        promptTokens:
          type: [integer, 'null']
          minimum: 0
        completionTokens:
          type: [integer, 'null']
          minimum: 0
        totalTokens:
          type: [integer, 'null']
          minimum: 0
        reasoningTokens:
          type: [integer, 'null']
          minimum: 0
        generatedAt:
          type: string
          format: date-time
    Report:
      type: object
      required:
        - id
        - auditId
        - source
        - architectureSummary
        - mermaidGraph
        - riskyAssumptions
        - acceptanceCriteria
        - promptPack
        - rolloutNotes
        - riskScore
        - verdict
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        auditId:
          type: string
          format: uuid
        source:
          type: string
        architectureSummary:
          type: string
        mermaidGraph:
          type: string
        riskyAssumptions:
          type: array
          items:
            $ref: '#/components/schemas/RiskyAssumption'
        acceptanceCriteria:
          type: array
          items:
            $ref: '#/components/schemas/AcceptanceCriterion'
        promptPack:
          $ref: '#/components/schemas/PromptPack'
        rolloutNotes:
          type: string
        smokeTestResults:
          oneOf:
            - $ref: '#/components/schemas/SmokeTestResults'
            - type: 'null'
        smokeScreenshotUrl:
          type: [string, 'null']
        analysisUsage:
          oneOf:
            - $ref: '#/components/schemas/AnalysisUsage'
            - type: 'null'
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        verdict:
          $ref: '#/components/schemas/Verdict'
        testPlan:
          oneOf:
            - $ref: '#/components/schemas/TestPlan'
            - type: 'null'
        affectedAreas:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/AffectedArea'
            - type: 'null'
        shareSlug:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    ShareLinkResponse:
      type: object
      required: [shareSlug, shareUrl]
      properties:
        shareSlug:
          type: string
        shareUrl:
          type: string
    RevokeShareLinkResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
...
    PublicReport:
      type: object
      required: [report, audit]
      properties:
        report:
          $ref: '#/components/schemas/Report'
        audit:
          $ref: '#/components/schemas/PublicAudit'
    PublicAudit:
      type: object
      required: [requestedChange, githubUrl, liveUrl, detectedFramework, detectedPackageManager, detectedDbLayer, detectedAuthLayer, createdAt]
      properties:
        requestedChange:
          type: string
        githubUrl:
          type: [string, 'null']
        liveUrl:
          type: [string, 'null']
        detectedFramework:
          type: [string, 'null']
        detectedPackageManager:
          type: [string, 'null']
        detectedDbLayer:
          type: [string, 'null']
        detectedAuthLayer:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...